move blog feed and change link
