+++ blog/views.py
@@ -79,3 +79,18 @@ class LatestCategoryFeed(Feed):
 
     def item_description(self, item):
         return item.body
+
+
+class BlogsFeed(Feed):
+    title = "Cos.io blog"
+    link = "/blog/feed/"
+    description = "Blogs of the Center for Open Science"
+
+    def items(self):
+        return BlogPage.objects.live().order_by('-date')
+
+    def item_title(self, item):
+        return item.title
+
+    def item_description(self, item):
+        return item.intro
